requirements.txt
# M1 and M2 requirements

# Core dependencies (install first)
python-dotenv==1.0.0
boto3==1.33.13
arxiv==2.1.0

# Data processing - install before ML packages
numpy==1.24.4
pandas==2.1.4

# PDF Processing
PyMuPDF==1.23.26
pdfplumber==0.10.3
pytesseract==0.3.10
camelot-py==0.11.0  # Table extraction from PDFs
# Note: If you need ghostscript support for camelot, install separately:
# macOS: brew install ghostscript
# Ubuntu: sudo apt-get install ghostscript
Pillow==10.2.0
opencv-python==4.8.1.78


# Testing & utilities
pytest==7.4.4
pytest-asyncio==0.23.3
tqdm==4.66.1

# Dependency conflict resolution
pyparsing>=3.1.0  # Required by pydot 4.0.1
PyYAML>=6.0  # Fix for build issues with newer setuptools (has pre-built wheels for Python 3.11)

# Pinecone (for vector search)
pinecone

# OpenAI (for embeddings and chat completions)
openai>=1.0.0
tiktoken>=0.5.0

# LangGraph (for multi-agent workflow)
langgraph>=0.0.20

# FastAPI (for M4 API)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
# email-validator>=2.0.0
